refactor: rename core/workflow/ -> core/workflow-engine/
Disambiguates the rule engine (state-machine, when-evaluator, timeouts)
from core/workflows/ (implementations like onboarding). Pure git mv +
import-path rewrite; docs updated.
