+++ core/__tests__/workflow/step-types.test.ts
@@ -18,7 +18,7 @@ import configManager from '../../infrastructure/config-manager'
 import pathManager from '../../infrastructure/path-manager'
 import { customWorkflowStorage } from '../../storage/custom-workflow-storage'
 import { workflowRuleStorage } from '../../storage/workflow-rule-storage'
-import { executeWorkflowRules } from '../../workflow/workflow-engine'
+import { executeWorkflowRules } from '../../workflow-engine/workflow-engine'
 
 async function freshProject(): Promise<{ projectPath: string; projectId: string }> {
   const projectPath = await fs.mkdtemp(path.join(os.tmpdir(), 'prjct-step-types-test-'))
