+++ core/commands/workflow/rule-actions/rules.ts
@@ -9,7 +9,7 @@ import type { CommandResult } from '../../../types/commands'
 import type { WorkflowRule } from '../../../types/storage/extended'
 import { failWith, notifyDone, notifyFail } from '../../../utils/md-aware'
 import { mdDone, mdList, mdNextSteps, mdOutput, mdSection } from '../../../utils/md-formatter'
-import { WORKFLOW_TIMEOUTS } from '../../../workflow/timeouts'
+import { WORKFLOW_TIMEOUTS } from '../../../workflow-engine/timeouts'
 import { requireWorkflow } from '../../guards'
 import { parseAction, searchRules } from '../intent'
 import { MAX_LISTED_MATCHES, newRuleDefaults, RULE_POSITIONS, type RulePosition } from './_shared'
