+++ core/services/task-service.ts
@@ -22,7 +22,7 @@ import type { CurrentTask, TaskFeedback } from '../schemas/state'
 import { getGitBranch } from '../session/git-helpers'
 import { stateStorage } from '../storage/state-storage'
 import * as dateHelper from '../utils/date-helper'
-import { executeWorkflowRules } from '../workflow/workflow-engine'
+import { executeWorkflowRules } from '../workflow-engine/workflow-engine'
 import { memoryService } from './memory-service'
 import projectService from './project-service'
 import { deriveWorkspace } from './workspace-id'
